Work with GameObjects in a 2D Scene
In this tutorial, you’ll:
Create a new Sprite.
Change a Sprite image using the Sprite Renderer component.
Select 2D objects in both the Hierarchy and Scene view windows of the Unity Editor.
Move, rotate, and scale 2D objects using numerical values using the Transform Component in the Inspector window.
Move, rotate, and scale 2D objects in the Scene view using the Unity transform tools.
Identify the relationship between the properties displayed in the Inspector and the display of GameObjects in the Scene.
Use the Rect Transform tool to manipulate a Sprite in the Scene.
Change a Sprite’s Tint colors.